Added basic Corpus object.
Implemented the basic corpus object, which encapsulates a mapping from
strings to unique integers, an integer array representing the raw
corpus data, and a suffix array which acts as an ordered index into the
corpus data. Added code to create a corpus object from a text file by
tokenizing the file, and also wrote code to create a suffix array for a
corpus based on Go's built in sort interface.