Improved the performance of the binary search algorithms and improved test coverage.
Improved the performance of the binary search over the suffix array by
implementing binary search methods that find the first and last
elements that match the required ngram. These methods also return a
lower / upper bound on where the opposite end of the range can be
located, and so the first call can be used to constrain the range of
the subsequent call. Also improved the test coverage.