Reimplemented the suffix array as a slice of slices.
Reimplemented the suffix array as a slice of slices into the corpus,
instead of using an integer slice as indexes into the corpus. This is
more idiomatic Go, and it also results simpler code and performance
improvements. Also added some additional tests and examples, as well as
code to perform some simple ngram language modeling.